=finish login,send order

+++ ProjectEBook/ebook_back/src/main/java/com/example/ebook_back/Dao/BookDao.java
@@ -20,8 +20,7 @@ import java.util.List;
 
 @Repository
 public interface  BookDao extends JpaRepository<Book, Integer>{
-
-    Book findBookById(long id);
+    Book findBookById(int id);
     List<Book> findAll();
 
 }
